requirements.txt
# to install run pip3 install -r requirements.txt
# to upgrade run pip3 install --upgrade -r requirements.txt

PySide6
Pillow
python-dotenv
google-generativeai
git+https://github.com/google-gemini/generative-ai-python@imagen